refactor: Isolate and fix panel drag logic
