+++ web/src/storeFunctions.ts
@@ -26,9 +26,16 @@ import {
   loadingCharts,
   selectedLayerStyles,
 } from "./store";
-import { getProjects, getDataset, getProjectCharts, getProjectSimulationTypes, getProjectDatasets, getDatasetLayers } from "@/api/rest";
+import {
+  getProjects,
+  getDataset,
+  getProjectCharts,
+  getProjectSimulationTypes,
+  getProjectDatasets,
+  getDatasetLayers,
+} from "@/api/rest";
 import { clearMapLayers, updateBaseLayer, updateLayersShown } from "./layers";
-import { Dataset, FloatingPanelConfig, Project } from "./types";
+import { Dataset, Project } from "./types";
 
 export function clearState() {
   clearProjectState();
@@ -81,96 +88,6 @@ export function clearProjectState() {
   availableSimulationTypes.value = undefined;
   currentSimulationType.value = undefined;
   selectedSourceRegions.value = [];
-}
-
-export function startDrag(
-  event: MouseEvent,
-  panel: FloatingPanelConfig | undefined,
-  modes: ("position" | "width" | "height")[]
-) {
-  if (panel) {
-    draggingPanel.value = panel.id;
-  }
-  draggingFrom.value = {
-    x: event.clientX,
-    y: event.clientY,
-  };
-  dragModes.value = modes;
-}
-
-export function dragPanel(event: MouseEvent) {
-  let offsetX = -5;
-  const offsetY = 30;
-  const snapToleranceX = 200;
-  const snapToleranceY = 300;
-  const minHeight = 100;
-  const minWidth = 150;
-
-  const panel = panelArrangement.value.find(
-    (p) => p.id === draggingPanel.value
-  );
-  if (!panel) return undefined;
-
-  if (panel.right) offsetX += document.body.clientWidth - 390;
-  const position = {
-    x: event.clientX - offsetX - (panel.element?.clientWidth || 0),
-    y: event.clientY - offsetY,
-  };
-  if (dragModes.value?.includes("position")) {
-    if (
-      panel.initialPosition &&
-      Math.abs(position.x - panel.initialPosition.x) < snapToleranceX &&
-      Math.abs(position.y - panel.initialPosition.y) < snapToleranceY
-    ) {
-      // snap to sidebar
-      panel.position = undefined;
-    } else {
-      if (!panel.initialPosition) {
-        // convert to floating
-        panel.width = 300;
-        panel.height = 200;
-        panel.initialPosition = position;
-      } else if (
-        panel.width && position.x + panel.width < document.body.clientWidth &&
-        panel.height && position.y + panel.height < document.body.clientHeight
-      ) {
-        panel.position = position;
-      }
-    }
-  }
-  if (draggingFrom.value) {
-    const from: { x: number; y: number } = { ...draggingFrom.value };
-    if (dragModes.value?.includes("height") && draggingFrom.value) {
-      if(!panel.height) {
-        panel.height = panel.element?.clientHeight
-      }
-      if (panel.height) {
-        const heightDelta = event.clientY - draggingFrom.value.y;
-        if (panel.height + heightDelta > minHeight) {
-          panel.height = panel.height + heightDelta;
-          from.y = event.clientY;
-        }
-      }
-    }
-    if (
-      dragModes.value?.includes("width") &&
-      draggingFrom.value &&
-      panel.width
-    ) {
-      const widthDelta = event.clientX - draggingFrom.value.x;
-      if (panel.width + widthDelta > minWidth) {
-        panel.width = panel.width + widthDelta;
-        from.x = event.clientX;
-      }
-    }
-    draggingFrom.value = from;
-  }
-}
-
-export function stopDrag() {
-  draggingPanel.value = undefined;
-  draggingFrom.value = undefined;
-  dragModes.value = [];
 }
 
 export function getMap() {
